booking_api/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
pytest==7.4.0
python-dotenv==1.0.0
marshmallow==3.20.1
uuid==1.30
